Develop

+++ ConventionManager/Controllers/ChatEventController.cs
@@ -232,8 +232,17 @@ namespace ConventionManager.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id, string fromWhere)
         {
             var chatEvent = await _context.ChatEvents.FindAsync(id);
-            _context.ChatEvents.Remove(chatEvent);
-            await _context.SaveChangesAsync();
+            IEnumerable<Subscription> subscriptions = _context.Subscriptions.Where(s => s.Event.Id == id).ToArray();
+            if (subscriptions.Any())
+            {
+                TempData["CannotDeleteEvent"] = chatEvent.CannotDeleteEventMessage;
+            }
+            else
+            {
+                _context.ChatEvents.Remove(chatEvent);
+                await _context.SaveChangesAsync();
+            }
+
             // Checks where the request came from to redirect correctly
             switch (fromWhere)
             {
