+++ ConventionManager/Controllers/PartyEventController.cs
@@ -229,8 +229,17 @@ namespace ConventionManager.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id, string fromWhere)
         {
             var partyEvent = await _context.PartyEvents.FindAsync(id);
-            _context.PartyEvents.Remove(partyEvent);
-            await _context.SaveChangesAsync();
+            IEnumerable<Subscription> subscriptions = _context.Subscriptions.Where(s => s.Event.Id == id).ToArray();
+            if (subscriptions.Any())
+            {
+                TempData["CannotDeleteEvent"] = partyEvent.CannotDeleteEventMessage;
+            }
+            else
+            {
+                _context.PartyEvents.Remove(partyEvent);
+                await _context.SaveChangesAsync();
+            }
+            
             // Checks where the request came from to redirect correctly
             switch (fromWhere)
             {
